bcc-site/requirements.txt
Flask>=1.0.2
PyMySQL>=0.9.2
flask-login
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pyopenssl